WIP: Implement unit tests for Skill construction
