+++ tests/skills/test_skill_inventory.py
@@ -1,0 +1,1 @@
+"""Unit tests for the SkillInventory class."""
